# forge-go/api/openapi/openapi.json
{
...
  "info": {
    "title": "Rustic AI API",
    "version": "0.0.9"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/oauth/callback": {
      "get": {
        "tags": [
          "oauth"
        ],
        "summary": "OAuth Callback",
        "operationId": "oauthCallback",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "error",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "An HTML page indicating success or failure.",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  }
}